pyproject.toml
[project]
name = "aceengineer-website"
version = "0.1.0"
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "pyyaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
